requirements.txt
django==3.2.16
django-environ==0.9.0
django-cors-headers==3.13.0
djangorestframework==3.14.0

argon2-cffi==21.3.0
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.5.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0

Pillow==9.3.0
requests==2.28.2
django-storages==1.13.1
django-ckeditor==6.3.2
psycopg2==2.9.5
python-slugify==6.1.2